# packages/api/openapi.yaml
openapi: "3.0.3"
info:
  title: "iQbe API"
  version: 2.1.0
servers:
  - url: "http://localhost:9000/api/v2"
...
security:
  - jwt: []
paths:
...
  "/invite-code":
    get:
      tags:
        - auth
      summary: 招待コード
      description: 招待コードのリストを取得
      parameters:
        - name: status
          in: query
          required: false
          description: '0: 未使用, 1: 使用済み'
          schema: 
            type: number
          example: 0
        - name: sort
          in: query
          required: false
          description: 'asc: 作成日時（昇順），desc: 作成日時（降順）'
          schema:
            type: string
          example: asc
      responses:
        "200":
            description: 招待コードのリストを取得
            content:
              application/json:
                schema: 
                  type: array
                  items:
                    type: object
                    properties:
                      code: 
                        type: string
                      status:
                        type: number
                      created: 
                        type: string
                        format: date-time
                    required:
                      - code
                      - status
                      - created
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
    post:
      tags: 
        - auth
      summary: 招待コードを発行
      description: 招待コードを発行する
      responses:
        "200":
          description: 招待コードを発行
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: 
                    type: string
                required:
                  - code
              example:
                code: 96892E7B-3B19-4BF9-96AA-5C55F4E8BCAC
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
...
  "/like":
    post: 
      tags: 
        - quizzes
      summary: お気に入りにクイズを登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                qid: 
                  type: string
              required:
                - qid
            example: 
              qid: DG8ED3bx7y0McbistuLx
      responses:
        "201":
          $ref: "#/components/responses/QuizResponse"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
  "/unlike":
    post: 
      tags:
        - quizzes
      summary: お気に入りからクイズを削除
      requestBody: 
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                qid: 
                  type: string
              required:
                - qid
            example: 
              qid: DG8ED3bx7y0McbistuLx
      responses:
        "204": 
          description: deleted quiz from favorite
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
...
  "/request-reset-password":
    post: 
      tags: 
        - auth
      summary: パスワードリセットを要求
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              type: object
              properties:
                username: 
                  type: string
                email: 
                  type: string
              required:
                - username
                - email
            example: 
              username: hoge
              email: hoge@example.com
      responses:
        "201": 
          description: 成功
          content:
            application/json:
              schema: 
                type: string
              example: "send email"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
  "/reset-password": 
    post: 
      tags:
        - auth
      summary: パスワードリセットを実行
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              type: object
              properties:
                token: 
                  type: string
                newPassword:
                  type: string
              required:
                - token
                - newPassword
            example: 
              token: 78FE38FF-8AC7-4625-BF54-D4695EB30EE2
              newPassword: newP@ssw0rd
      responses:
        "200":
          description: リセット成功
          content:
            application/json:
              schema: 
                type: string
              example: "succeed reset password"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
...
  "/workbooks":
    get: 
      tags:
        - workbooks
      summary: 自分が作った問題集のリストを取得
      responses:
        "200":
          $ref: "#/components/responses/WorkbooksResponse"
    post:
      tags:
        - workbooks
      summary: 問題集を新規作成
      requestBody:
        required: true
        content:
          application/json:
            schema:  
              type: object
              properties:
                name:
                  type: string
                published: 
                  type: string
                  format: date-time
              required:
                - name
            example: 
              name: abc the 12th
      responses:
        "200":
          $ref: "#/components/responses/WorkbookResponse"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
  "/workbooks/{wid}":
    get: 
      tags:
        - workbooks
      summary: 問題集の詳細を取得
      description: リクエストパラメータに指定した問題集の詳細を取得します．
      parameters:
        - name: wid
          in: path
          required: true
          description: 問題集のID
          schema:
            type: string
          example: 1i7qaa
      responses:
        "200":
          $ref: "#/components/responses/WorkbookResponse"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
    put:  
      tags:
        - workbooks
      summary: 問題集の編集
      description: リクエストボディに指定した問題集を編集します．
      parameters:
        - name: wid
          in: path
          required: true
          description: 問題集のID
          schema:
            type: string
          example: 1i7qaa
      requestBody:
        required: true
        content:
          application/json:
            schema:  
              type: object
              properties:
                name:
                  type: string
                published: 
                  type: string
                  format: date-time
              required:
                - name
            example: 
              name: abc the 12th
      responses:
        "200":
          $ref: "#/components/responses/WorkbookResponse"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
    delete:
      tags: 
        - workbooks
      summary: 問題集を削除
      description: リクエストボディに指定した問題集を削除します．削除する問題集に登録されたクイズは削除されません．
      parameters:
        - name: wid
          in: path
          required: true
          description: 問題集のID
          schema:
            type: string
          example: 1i7qaa
      responses:
        "200":
          $ref: "#/components/responses/WorkbooksResponse"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
  "/workbooks/all":
    get:
      tags:
        - workbooks
      summary: 取得できるすべての問題集のリストを取得
      description: 公開された問題集，閲覧権限のある問題集，そして自分が作った問題集のリストを取得します．
      responses:
        "200":
          $ref: "#/components/responses/WorkbooksResponse"
        "401": 
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
components:
...
  responses:
    400BadRequest:
      description: パラメータの不備
      content:
        application/problem+json:
          schema: 
            $ref: "#/components/schemas/Error"
          examples:
            Invaild parameter: 
              value:
                title: INVAILD_PARAMS
                type: "about:blank"
                status: 400
                detail: "This request contains invaild parameter. Please correct to the right ones."
    401Unauthorized: 
      description: JWTトークンの不備（期限切れ，無効，存在しない）
      content:
        application/problem+json:
          schema: 
            $ref: "#/components/schemas/Error"
          examples: 
            Expired token:
              value:
                title: EXPIRED_TOKEN
                type: "about:blank"
                status: 401
                detail: "This authorization token have expired. Please refetch token."
            Invalid token:
              value:
                code: INVAILD_TOKEN
                type: "about:blank"
                status: 401
                detail: "This authorization token is invalid."
            No token:
              value:
                code: NoToken
                type: "about:blank"
                status: 401
                message: "Authorization token was not found. Please set authorization header."
    403Forbidden: 
      description: アクセス権限がない
      content:
        application/problem+json:
          schema: 
            $ref: "#/components/schemas/Error"
          examples:
            Access denied:
              value: 
                title: ACCESS_DENIED
                type: "about:blank"
                status: 403
                detail: "Access denied. Insufficient rights to access the resource."
...
    QuizResponse:
      description: クイズ情報
      content:
        application/json:
          schema: 
            $ref: "#/components/schemas/Quiz"
          example:
            id: 1
            question: すい臓にある組織・ランゲルハンス島の発見者であるランゲルハンスは、現在でいうどこの国の病理学者でしょう？
            answer: ドイツ
            workbook: abc the 12th
            wid: sX8G8W
            level: 1
            category: 10
            subCategory: 14
            total: 10
            right: 4
            isFavorite: false
            registerdMylist: []
            creatorId: 1
            isPublic: false
...
    WorkbookResponse: 
      description: 問題集のリスト
      content:
        application/json:
          schema: 
            $ref: "#/components/schemas/Workbook"
          example: 
            wid: sX8G8W
            name: abc the 12th
            creatorId: 1
    WorkbooksResponse: 
      description: 問題集のリスト
      content:
        application/json:
          schema: 
            type: array
            items:
              $ref: "#/components/schemas/Workbook"
          example: 
            - wid: sX8G8W
              name: abc the 12th
              creatorId: 1
            - wid: Kb5F7y
              name: abc the tenth
              creatorId: 1
  schemas:
    Category: 
      type: object
      properties:
        id: 
          type: number
        name: 
          type: string
        description:
          type: string
        disabled:
          type: boolean
        from: 
          type: string
        parentId:
          type: number
        sub: 
          type: array
          items:
            $ref: "#/components/schemas/SubCategory"
      required:
        - id
        - name
        - disabled
        - parentId
    Error: 
      type: object
      properties: 
        title:
          type: string
          description: エラーを簡潔に表現したコード
        type: 
          type: string
          description: エラーの詳細ドキュメントを指すURL
        status:
          type: number
          description: HTTPステータスコード
        detail: 
          type: string
          description: 開発者向けの詳細文
      required:
        - title
        - type
        - status
        - detail
...
    Mylist: 
      type: object
      properties:
        mid: 
          type: string
        name: 
          type: string
        created:
          type: string
          format: date-time
          readOnly: true
      required:
        - mid
        - name
        - created
    Quiz: 
      type: object
      properties:
        qid: 
          type: string
        question:
          type: string
        answer: 
          type: string
        anotherAnswer:
          type: string
        wid: 
          type: string
          writeOnly: true
        workbook:
          $ref: "#/components/schemas/Workbook"
          readOnly: true
        categoryId:
          type: number
          writeOnly: true
        category:
          type: array
          items:
            $ref: "#/components/schemas/Category"
          readOnly: true
        total: 
          type: number
          readOnly: true
        right: 
          type: number
          readOnly: true
        tagLabels:
          type: array
          items: 
            type: string
          writeOnly: true
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
          readOnly: true
        creatorId:
          type: string
        isPublic: 
          type: boolean
        limitedUser:
          type: array
          items:
            type: number
          writeOnly: true
        isFavorite:
          type: boolean
          readOnly: true
        registerdMylist:
          type: array 
          items: 
            $ref: "#/components/schemas/Mylist"
          readOnly: true
        judgement:
          type: number
          readOnly: true
      required:
        - qid
        - question
        - answer
        - total
        - right
        - creatorId
        - isPublic
        - isFavorite
...
    SubCategory: 
      type: object
      properties:
        id: 
          type: number
        name: 
          type: string
        description: 
          type: string
        parentId:
          type: number
        disabled:
          type: boolean
        from: 
          type: string
      required:
        - id
        - name
        - parentId
        - disabled
    Tag: 
      type: object
      properties:
        label: 
          type: string
        created: 
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
      required:
        - label
        - created
        - modified
...
    Workbook:
      type: object
      properties:
        wid: 
          type: string
        name: 
          type: string
        date: 
          type: string
          format: date-time
        levelId: 
          type: number
        color: 
          type: string
        creatorId:
          type: string
      required:
        - wid
        - name
        - creatorId
  securitySchemes:
    jwt:
      type: http
      description: Credentials or access token for API
      scheme: bearer
      bearerFormat: JWT